diabetes_blood_image_project/requirements.txt
streamlit
numpy
pandas
scikit-learn
opencv-python
pillow
joblib
matplotlib
